refactor(commands): move to outer scope
