+++ src/commands/apply.command.ts
@@ -10,7 +10,7 @@ import {
   SlashCommandBuilder,
 } from 'discord.js'
 
-import { Command } from './types'
+import { Command } from './types/index'
 
 import { Color, Image } from '~/config'
