+++ src/commands/ping.command.ts
